requirements.txt
# MediSync - Offline Drug Interaction Checker

# Data & ML
kagglehub>=0.2.0
pandas>=2.0.0

# Backend - FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing
httpx>=0.26.0
pytest>=7.0.0
